minor: addressing comments
