+++ backend/danswer/db/chat.py
@@ -1,4 +1,3 @@
-import datetime
 from uuid import UUID
 
 from sqlalchemy import delete
@@ -156,53 +155,6 @@ def delete_chat_session(
     db_session.commit()
 
 
-class ChatSessionSkeleton:
-    session_id: int
-    user_id: int
-    one_shot: bool
-    time_created: datetime.datetime
-    time_updated: datetime.datetime
-
-    def __init__(
-        self,
-        session_id: int,
-        user_id: int,
-        one_shot: bool,
-        time_created: datetime.datetime,
-        time_updated: datetime.datetime,
-    ) -> None:
-        self.session_id = session_id
-        self.user_id = user_id
-        self.one_shot = one_shot
-        self.time_created = time_created
-        self.time_updated = time_updated
-
-
-def get_chat_sessions_skeleton(
-    db_session: Session,
-) -> list[ChatSessionSkeleton]:
-    stmt = select(
-        ChatSession.id,
-        ChatSession.user_id,
-        ChatSession.one_shot,
-        ChatSession.time_created,
-        ChatSession.time_updated,
-    )
-
-    result = db_session.execute(stmt).all()
-
-    return [
-        ChatSessionSkeleton(
-            session_id=s.id,
-            user_id=s.user_id,
-            one_shot=s.one_shot,
-            time_created=s.time_created,
-            time_updated=s.time_updated,
-        )
-        for s in result
-    ]
-
-
 def get_chat_message(
     chat_message_id: int,
     user_id: UUID | None,
@@ -248,40 +200,6 @@ def get_chat_messages_by_session(
     result = db_session.execute(stmt).scalars().all()
 
     return list(result)
-
-
-class ChatMessageSkeleton:
-    message_id: int
-    chat_session_id: int
-    time_sent: datetime.datetime
-
-    def __init__(
-        self, message_id: int, chat_session_id: int, time_sent: datetime.datetime
-    ) -> None:
-        self.message_id = message_id
-        self.chat_session_id = chat_session_id
-        self.time_sent = time_sent
-
-
-# Gets skeletons of all message
-# TODO: should change this to use a key index with dates
-# TODO: Need to paginate as well
-def get_empty_chat_messages_entries(
-    db_session: Session,
-) -> list[ChatMessageSkeleton]:
-    stmt = select(
-        ChatMessage.id, ChatMessage.chat_session_id, ChatMessage.time_sent
-    ).where(ChatMessage.message_type == MessageType.USER)
-
-    result = db_session.execute(stmt).all()
-
-    print(result)
-    return [
-        ChatMessageSkeleton(
-            message_id=m.id, chat_session_id=m.chat_session_id, time_sent=m.time_sent
-        )
-        for m in result
-    ]
 
 
 def get_or_create_root_message(
